Add files via upload
College name: Velammal College Of Engineering And Technology
Department: Computer Science Engineering
Guide name: Dr.S.Sasikala,Phd.,/Associate Professor
Team details: Miruthula N, Poornimadevi A, Suuky M, Subadhina J
Artificial neural network (ANN) plays an important role in many medical imaging applications. The detection of cervical cancer cells uses an ANN for classifying the normal and abnormal cells in the cervix region of the uterus. Cervical cancer detection is very challenging because this cancer occurs without any symptoms. The classification between the normal, abnormal and cancerous cells is identified by using an artificial neural network which produces accurate results than the manual screening methods like Pap smear and Liquid cytology based (LCB) test. The ANN uses several architectures for easy and accurate detection of cervical cells. In this project, a survey and analysis on the different types of architecture in the ANN with its accuracy results and performance are discussed. A brief description about the working and detection of cervical cancer is presented which is useful for the classification of normal and abnormal cervical cells.